feat(queue): implement queue_append
Operation should be fairly quick since we keep track of the last element
of the list using a pointer.

Future work could be to use a double linked list instead, this way we
could decide to search the list from the back which could speed up
queue_remove for higher IDs. Probably unnecessary, since we expect to
delete IDs near the front mostly.