feat(bahnhofshalle): migrate to parcel for building
We get

* rid of redo
* minified assets
* output hashes (for aggressive caching)

Also we use node2nix instead of yarn2nix. A node_modules directory which
contains the entire dependency tree is something required for parcel to
work. Unfortunately this is not provided currently by yarn2nix (I'll
work on it again sometime). In order to not clutter up the build process
and increase the build time with an experimental pinned version of
yarn2nix, we'll just use node2nix.