refactor(warteraum): replace json_output with own emitjson
Still somewhat limited, but implements everything we need in warteraum.
API is very similar to json_output, but we choose to not implement
buffer handling and use open_memstream instead. The FILE * abstraction
suits us very well here since we just want to write chars to some kind
of output.