chore(warteraum): port build system to GNU make
make is what people already use and it integrates better with nix, so
lets just use it as well. I haven't yet figured out how to do cleanly
implement a build system with redo without a thousand annoying little
shell scripts.

The downside of this is that GNU make is sometimes too lazy when
incrementally building and doesn't rebuild stuff when it really should.