feat(warteraum): read tokens and salt from files
Instead of compiling salt and tokens in, read both from files that are
specified as environment variables WARTERAUM_SALT_FILE and
WARTERAUM_TOKENS_FILE.

hashtoken has also been adjusted to read the salt from a specified salt
file (the first argument). It also now outputs the token in raw form
instead of C syntax. Intended usage is hashtoken
`/path/to/salt token >> /path/to/tokens`. Disadavantage of this is that
deleting tokens is somewhat cumbersome and only really doable with a hex
editor.

The NixOS service now expects saltFile and tokensFile instead of salt and
tokens as a string and a list of strings respectively.