fix(nixos): fix sandboxing
Due to a typo, previously the filesystem was accessible. Now we need
can't use `InaccessiblePaths` anymore, since it doesn't allow nesting
with `BindReadOnlyPaths`. Thus we need to switch to
`TemporaryFileSystem` which unfortunately doesn't seem to work with
`DynamicUser`.

With a new warteraum-specific user we can run warteraum in a
filesystem that only contains `/nix/store`, the secret files and vital
things like `/proc`, `/dev` etc.