refactor(bahnhofshalle): use esbuild over parcel-bundler
parcel-bundler is a bit of an annoyance, as it has a million of
dependencies (in typical js fashion) and is not built by NixOS's Hydra.
An alternative to parcel-bundler is esbuild which is written in Go.

It, however, doesn't support transpiling ES6 to ES5 which we'll accept
for the lessened deployment annoyances. All modern browsers support ES6
anyways (especially the subset we are using).