doc(bahnhofshalle): document JS requirement in FAQ
Since we are cross-referencing an FAQ section, we also make this
convenient by adding ids for all headings and highlighting the currently
targeted heading with an underline. This ensurse that it's clear which
section to read when clicking on a link even when the screen may not
jump around on even normal sized screens.