chore(warteraum): remove support for v1 API
I could not find any evidence anyone has been using this API this
year (except some AI companies' crawler bots sending GET requests to the
wrong endpoints), so it seems like it's as good a time as any to pull
the plug.

The old endpoints now return 410 Gone. bahnhofshalle without JavaScript
will now use the v2 API as a fallback, meaning that the user will be
shown JSON instead of a (lackluster) HTML page.

Due to the whitespace changes, it is easiest to review this change with
`git diff -b`.